bin: Add var for vm count, the amount of vms in jem_env.vms array, issue #9
Previously, this was done incorrectly by checking the size of the array,
which falsely returned the size of the pointer to the array. Rather than,
the actual array size, which is now being tracked in jem_env.vm_count
variable. This fixes the issue with jem not being able to set vm by numeric
value beyond 3; more than 3 vms and trying to set the 4th+ via numeric value.

Fixes #9
https://github.com/Obsidian-StudiosInc/jem/issues/9